// src/features/metronome/enums/TimeSignature.ts
export enum TimeSignature {
  OneFour = '1 / 4',
  TwoFour = '2 / 4',
  ThreeFour = '3 / 4',
  FourFour = '4 / 4',
  FiveFour = '5 / 4',
  SixFour = '6 / 4',
  SevenFour = '7 / 4',
  EightFour = '8 / 4',
  NineFour = '9 / 4',
  TenFour = '10 / 4',
  ElevenFour = '11 / 4',
  TwelveFour = '12 / 4',
  OneTwo = '1 / 2',
  TwoTwo = '2 / 2',
  FourTwo = '4 / 2',
  ThreeEight = '3 / 8',
  SixEight = '6 / 8',
  SevenEight = '7 / 8',
  EightEight = '8 / 8',
  NineEight = '9 / 8',
  TenEight = '10 / 8',
  ElevenEight = '11 / 8',
  TwelveEight = '12 / 8',
}